implemented Mapstruct instead of manual mappers, fixed code reviev, added method for calculate budget

+++ database/create_table_projects.sql
@@ -2,9 +2,9 @@ drop table if exists projects;
 create table projects
 (
     id           bigserial primary key,
-    project_name varchar(255) not null unique,
+    project_name varchar(255)   not null unique,
     description  varchar(255),
-    start_date   date         not null check (start_date < end_date),
-    end_date     date         not null check (end_date > start_date),
-    budget       float        not null check (budget > 0)
+    start_date   date           not null check (start_date < end_date),
+    end_date     date           not null check (end_date > start_date),
+    budget       numeric(20, 2) not null check (budget > 0)
 )
